buffalo_package/requirements.txt
pyyaml>=5.1
setuptools>=42.0.0 